Refactor user management to use Usuarios model and secure password hashing
Replaces legacy Usuario model with Usuarios and UsuariosEXT across business and web layers. Adds PasswordHasher for secure password storage and verification using PBKDF2. Updates CRUD operations, user-related stored procedure calls, and refactors usages in web pages and controls. Changes user ID types from int to long for consistency. Improves security and maintainability of user authentication and management.

+++ BLS.Negocio/Expediente/DatosExpedientes.cs
@@ -382,7 +382,7 @@ namespace BLS.Negocio.Expediente
             }
         }
 
-        public List<ListaExpedientes> DameListaExpediente(DateTime fechaInicial, DateTime fechaFinal,  int idUsuario, Boolean todasLasFechas)
+        public List<ListaExpedientes> DameListaExpediente(DateTime fechaInicial, DateTime fechaFinal,  long idUsuario, Boolean todasLasFechas)
         {
             try
             {
